requirements.txt
engineering-notation==0.6.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
Pillow==8.1.0
tk-tools==0.13.0
